Implemented pause button
p key now pauses game and displays pause message. The pause message seems to display properly most of the time, but it will weirdly have errors at random runs or compilations. If the messages doesn't appear properly, try compiling or running it again. Regardless of whether the message displays properly or not, all controls are disabled when the game is paused.